% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/data2007.R
\docType{data}
\name{d}
\alias{d}
\alias{d}
\title{Example data from HW 1}
\format{An object of class \code{data.frame} with 30 rows and 2 columns.}
\usage{
d

d
}
\description{
Example data from HW 1

Example data from lecture 6
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/q1p4.txt}

\url{http://users.stat.umn.edu/~almquist/3811_examples/gapminder2007ex.csv}
}
\author{
Charlie Geyer \email{geyer@umn.edu}

Zack almquist \email{almquist@umn.edu}
}
\keyword{datasets}
