% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JiangZhihangToolsMLEfunc.R
\name{MLEfunc}
\alias{MLEfunc}
\title{MLE}
\usage{
MLEfunc(x, f, interval)
}
\arguments{
\item{x}{vector}

\item{interval}{vector, i.e., interval for optimize function}

\item{func}{function, e.g., `function(theta, x) dgamma(x, shape = theta, log = TRUE)`}
}
\value{
scalar
}
\description{
Computes the liklihood of a given distribution for data x
}
\examples{
x1=rgamma(100,3)
f = function(theta, x) dgamma(x, shape = theta, log = TRUE)
result_gamma <- MLEfunc(x1,f,c(0,3))
result_gamma
}
